+++ core-bundle/src/Resources/contao/controllers/BackendPreview.php
@@ -102,7 +102,7 @@ class BackendPreview extends \Backend
 							   ->execute($objUser->id, time(), 'FE_USER_AUTH', \System::getContainer()->get('session')->getId(), \Environment::get('ip'), $strHash);
 
 				// Set the cookie
-				$this->setCookie('FE_USER_AUTH', $strHash, (time() + \Config::get('sessionTimeout')), null, null, false, true);
+				$this->setCookie('FE_USER_AUTH', $strHash, (time() + \Config::get('sessionTimeout')), null, null, \Environment::get('ssl'), true);
 				$objTemplate->user = \Input::post('user');
 			}
 		}