+++ core-bundle/src/Resources/contao/controllers/BackendSwitch.php
@@ -84,14 +84,14 @@ class BackendSwitch extends \Backend
 			// Hide unpublished elements
 			if (\Input::post('unpublished') == 'hide')
 			{
-				$this->setCookie('FE_PREVIEW', 0, ($time - 86400));
+				$this->setCookie('FE_PREVIEW', 0, ($time - 86400), null, null, \Environment::get('ssl'), true);
 				$objTemplate->show = 0;
 			}
 
 			// Show unpublished elements
 			else
 			{
-				$this->setCookie('FE_PREVIEW', 1, ($time + \Config::get('sessionTimeout')));
+				$this->setCookie('FE_PREVIEW', 1, ($time + \Config::get('sessionTimeout')), null, null, \Environment::get('ssl'), true);
 				$objTemplate->show = 1;
 			}
 
@@ -114,7 +114,7 @@ class BackendSwitch extends \Backend
 									   ->execute($objUser->id, $time, 'FE_USER_AUTH', \System::getContainer()->get('session')->getId(), \Environment::get('ip'), $strHash);
 
 						// Set the cookie
-						$this->setCookie('FE_USER_AUTH', $strHash, ($time + \Config::get('sessionTimeout')), null, null, false, true);
+						$this->setCookie('FE_USER_AUTH', $strHash, ($time + \Config::get('sessionTimeout')), null, null, \Environment::get('ssl'), true);
 						$objTemplate->user = \Input::post('user');
 					}
 				}
@@ -123,7 +123,7 @@ class BackendSwitch extends \Backend
 				else
 				{
 					// Remove cookie
-					$this->setCookie('FE_USER_AUTH', $strHash, ($time - 86400), null, null, false, true);
+					$this->setCookie('FE_USER_AUTH', $strHash, ($time - 86400), null, null, \Environment::get('ssl'), true);
 					$objTemplate->user = '';
 				}
 			}